Code Restructure To Reduce File Usage and Help Readability
- Move from github-submodule structure to use CMake's native dependency
  management facilities.
  (https://cmake.org/cmake/help/latest/guide/using-dependencies/index.html)

- Restructure The #include statements to indicate the project source,
  this prevents confusion and is easier to read. It is also a consistent
  way to describe the included header and to reason about the include
  structure.

- Explicit Code file declaration in CMake as opposed to File Globbing
  which is only run once during build system generation. When creating a
  new file, it is suggested to specify it in the CMake File which will
  enhance tracking and ensure that it is picked up at build time due to
  changes in the CMake Files.

- Changed Bitmap.h to utf-8 from iso-8859

- Changed from MSVC style of pre-compiled headers, just using plain
  header includes for portability. Another benefit is storage savings.

+++ ZEngine/ZEngine/Windows/Events/KeyEvent.h
@@ -1,7 +1,7 @@
 #pragma once
-#include <Core/CoreEvent.h>
-#include <Inputs/KeyCode.h>
-#include <ZEngineDef.h>
+#include <ZEngine/Core/CoreEvent.h>
+#include <ZEngine/Windows/Inputs/KeyCode.h>
+#include <ZEngine/ZEngineDef.h>
 
 namespace ZEngine::Windows::Events
 {
@@ -41,7 +41,7 @@ namespace ZEngine::Windows::Events
 
         virtual std::string ToString() const override
         {
-            return fmt::format("KeyPressedEvent : {0}, repeated count : {1}", m_keycode, m_repeat_count);
+            return fmt::format("KeyPressedEvent : {0}, repeated count : {1}", static_cast<int32_t>(m_keycode), m_repeat_count);
         }
 
     protected:
@@ -67,7 +67,7 @@ namespace ZEngine::Windows::Events
 
         virtual std::string ToString() const override
         {
-            return fmt::format("KeyReleasedEvent : {0}", m_keycode);
+            return fmt::format("KeyReleasedEvent : {0}", static_cast<int32_t>(m_keycode));
         }
     };
 } // namespace ZEngine::Windows::Events
